moviendo todo lo referido a has_parameters.rb
