--- src/conditions/proc_parameters.rb
@@ -1,5 +1,0 @@
-class ProcParameters
-  def filter(parameters, quantity, rule)
-    parameters.select { |parameter| rule.call parameter }.size.equal? quantity
-  end
-end
