--- src/conditions/quantity_parameters.rb
@@ -1,5 +1,0 @@
-class QuantityParameters
-  def filter(parameters,quantity,rule=nil)
-    parameters.size.equal?(quantity)
-  end
-end
