--- src/conditions/regexp_parameters.rb
@@ -1,7 +1,0 @@
-class RegexpParameters
-
-  def filter(parameters,quantity,regex)
-    parameters.map{|arg| arg[1].to_s}.grep(regex).size == quantity
-  end
-
-end
